Update README.md
I've made some changes to the readme file. Please peruse and make further changes or state the changes and I'll make them.

I added contributors but I don't know the names of people that have contributed to your project in any way. It'll be good to give them credits in the readme file.

When the privacy policy I'll create is merged, the link can be added.